Move getting and merging data to a new file
This is necessary to get the data as a service and not do it with every
page request.
